+++ src/Inciter/ALECG.cpp
@@ -194,7 +194,7 @@ ALECG::queryBnd()
     for (const auto& bndry : timedep) {
       std::unordered_set< std::size_t > nodes;
       for (const auto& s : bndry.template get< tag::sideset >()) {
-        auto k = m_bnode.find(s);
+        auto k = m_bnode.find(static_cast<int>(s));
         if (k != end(m_bnode)) {
           for (auto g : k->second) {      // global node ids on side set
             nodes.insert( tk::cref_find(d->Lid(),g) );
@@ -242,7 +242,7 @@ ALECG::norm()
   // Query nodes at which mesh velocity symmetry BCs are specified
   std::unordered_map<int, std::unordered_set< std::size_t >> ms;
   for (const auto& s : g_inputdeck.get< tag::ale, tag::symmetry >()) {
-    auto k = m_bface.find(s);
+    auto k = m_bface.find(static_cast<int>(s));
     if (k != end(m_bface)) {
       auto& n = ms[ k->first ];
       for (auto f : k->second) {
