+++ src/PDE/MultiMat/Problem/FieldOutput.cpp
@@ -251,7 +251,7 @@ MultiMatSurfOutput(
   // extract field output along side sets requested
   for (auto s : g_inputdeck.get< tag::field_output, tag::sideset >()) {
     // get face list for side set requested
-    auto b = bface.find(s);
+    auto b = bface.find(static_cast<int>(s));
     if (b == end(bface)) continue;
     const auto& faces = b->second;
     std::vector< tk::real > surfaceSol( faces.size() );
